rename EngineTestCase to JsonRpcTestCase
